Add solution for 2020 d2 p2 in java and test for it

+++ 2020/2020-2/src/main/java/DayTwo.java
@@ -10,7 +10,7 @@ public class DayTwo {
     public static void main(String[] args) throws URISyntaxException, IOException {
         Path path = Paths.get(Objects.requireNonNull(DayTwo.class.getClassLoader().getResource("input.data")).toURI());
 
-        long[] solution = new Puzzle(Files.readAllLines(path)).solve();
+        int[] solution = new Puzzle(Files.readAllLines(path)).solve();
         System.out.println(Arrays.toString(solution));
     }
 }
